requirements.txt
openmesh
sklearn
numpy>=1.16.2
torch==1.7.0
torchgeometry>=0.1.2
PyYAML==5.1.1
configargparse
smplx
pillow
opencv-python
tqdm
pyrender==0.1.36
trimesh>=2.37.6
scipy
open3d
times
DateTime
tensorboard
pandas
ipdb

